refactor(core): share one poll backoff, and get a true VM-creation number
Three more fixed 50ms polls were quantizing the same way the readiness poll
was: the supervisor PID-file wait inside the HVF driver's boot, the guest-agent
wait on the standby path, and wait_for_pid_exit on teardown. All four sites now
share mvm_core::poll_backoff rather than four copies of a constant.

HVF, 20 runs plus 2 warm-ups, p50:

                        flat tick   readiness fixed   all four fixed
  driver_boot             53.8ms         53.8ms            4.6ms
  guest_kernel_entry      53.8ms         18.0ms           65.1ms
  VMM + guest boot       107.6ms         71.8ms           69.7ms
  dispatch window        117.2ms         81.4ms           79.2ms
  total                       -         352.1ms          310.1ms

Stated honestly: the readiness fix was worth a real 36ms; this one is worth
about 2ms of wall clock. What it buys is a true number. VM creation on this
backend costs 4.6ms, not the 53.8ms the tick reported, and the ~65ms is
genuinely the guest booting to a serving agent. The work did not leave the
launch, it moved to the span that was always doing it.

It also widened the tail, dispatch p99 83.5ms to 156.7ms. The flat tick had
been rounding every launch up to the same value and hiding the variance; the
distribution was never as tight as it looked. A p99 that is tight because of
quantization is not a better launch.

Two consequences. Phase 3's target on HVF is now known to be guest boot rather
than VMM setup — 65 of the 70ms. And Firecracker's 623.6ms driver_boot should
be re-measured against these fixes before being decomposed, since it contains
a poll of its own.
