+++ crates/mvm-core/src/lib.rs
@@ -60,6 +60,8 @@ pub mod kernel_format;
 pub mod launch_metadata;
 /// Backend-recorded launch phases, so a caller can see inside `start`.
 pub mod launch_trace;
+/// Shared backoff for polls that wait on a condition.
+pub mod poll_backoff;
 // Guest lifecycle markers + snapshot timing (the `mvm-init` ↔ host
 // contract) are a pure-DTO leaf that now lives in `mvm-contract`;
 // re-exported here as a module alias so every existing
